refactor(toast): defineOptions instead of two script once more
