+++ packages/nutui/components/toast/index.ts
@@ -1,2 +1,2 @@
-export * from './type'
+export * from './types'
 export * from './toast'
